moved sub-cache code to a separate lib
