+++ src/app/contextBroker/CMakeLists.txt
@@ -40,6 +40,7 @@ SET (STATIC_LIBS
     ngsi9
     ngsi10
     ngsi
+    cache
     mongoBackend
     parse
     orionTypes
