+++ src/lib/logMsg/traceLevels.h
@@ -87,9 +87,6 @@ typedef enum TraceLevels
 
   /* MongoBackend (100-119) */
   LmtMongo = 100,
-  LmtMongoSubCache,
-  LmtMongoSubCacheMatch,
-  LmtMongoCacheSync,
 
   /* Cleanup (120-139) */
   LmtDestructor = 120,
@@ -119,8 +116,13 @@ typedef enum TraceLevels
   LmtCacheSem,
   LmtTimeStatSem,
 
-  /* Others (>=210) */
-  LmtCm = 210,
+  /* Cache (210 - 230) */
+  LmtSubCache = 210,
+  LmtSubCacheMatch,
+  LmtCacheSync,
+
+  /* Others (>=230) */
+  LmtCm = 230,
   LmtIotaXmlReg,
   LmtIotaXmlObs,
   LmtRush,
