+++ src/lib/mongoBackend/MongoGlobal.cpp
@@ -76,6 +76,18 @@ static Notifier*            notifier;
 static bool                 multitenant;
 
 
+
+/* ****************************************************************************
+*
+* mongoMultitenant - 
+*/
+bool mongoMultitenant(void)
+{
+  return multitenant;
+}
+
+
+
 /* ****************************************************************************
 *
 * shutdownClient -
