+++ src/lib/mongoBackend/mongoSubCache.cpp
@@ -22,25 +22,6 @@
 *
 * Author: Ken Zangelin
 */
-
-/*
-*
-* The mongo subscription cache maintains in memory all subscriptions of type ONCHANGE that has
-* a patterned Entity::id. The reason for this cache is to avoid a very slow mongo operation ... (Fermin)
-*
-* The cache is implementged in mongoBackend/mongoSubCache.cpp/h and used in:
-*   - MongoCommonUpdate.cpp               (function addTriggeredSubscriptions_withCache)
-*   - mongoSubscribeContext.cpp           (function mongoSubscribeContext)
-*   - mongoUnsubscribeContext.cpp         (function mongoUnsubscribeContext)
-*   - mongoUpdateContextSubscription.cpp  (function mongoUpdateContextSubscription)
-*   - mongoSubCache.cpp                   (function mongoSubCacheRefresh)
-*
-* To manipulate the subscription cache, a semaphore is necessary, as various threads can be
-* reading/inserting/removing subs at the same time.
-* This semaphore is NOT optional, like the mongo request semaphore.
-*
-* Two functions have been added to common/sem.cpp|h for this: cacheSemTake/Give.
-*/
 #include <string>
 #include <regex.h>
 
@@ -52,581 +33,14 @@
 #include "common/sem.h"
 #include "common/string.h"
 
+#include "cache/subCache.h"
 #include "mongoBackend/MongoGlobal.h"
 #include "mongoBackend/connectionOperations.h"
 #include "mongoBackend/mongoSubCache.h"
 #include "mongoBackend/safeMongo.h"
 #include "mongoBackend/dbConstants.h"
 
-#include "ngsi/NotifyConditionVector.h"
-#include "ngsi10/SubscribeContextRequest.h"
-
 using namespace mongo;
-using std::map;
-
-
-
-/* ****************************************************************************
-*
-* EntityInfo::EntityInfo - 
-*/
-EntityInfo::EntityInfo(const std::string& _entityId, const std::string& _entityType, const std::string& _isPattern)
-{
-  entityId     = _entityId;
-  entityType   = _entityType;
-  isPattern    = (_isPattern == "true") || (_isPattern == "TRUE") || (_isPattern == "True");
-
-  if (isPattern)
-  {
-    regcomp(&entityIdPattern, _entityId.c_str(), 0);
-    entityIdPatternToBeFreed = true;
-  }
-  else
-  {
-    entityIdPatternToBeFreed = false;
-  }
-}
-
-
-
-/* ****************************************************************************
-*
-* EntityInfo::match - 
-*/
-bool EntityInfo::match
-(
-  const std::string&  id,
-  const std::string&  type
-)
-{
-  //
-  // If type non-empty - perfect match is mandatory
-  // If type is empty - always matches
-  //
-  if ((type != "") && (entityType != type) && (entityType != ""))
-  {
-    return false;
-  }
-
-  if (isPattern)
-  {
-    // REGEX-comparison this->entityIdPattern VS id
-    return regexec(&entityIdPattern, id.c_str(), 0, NULL, 0) == 0;
-  }
-
-  return (id == entityId);
-}
-
-
-
-/* ****************************************************************************
-*
-* EntityInfo::release - 
-*/
-void EntityInfo::release(void)
-{
-  if (entityIdPatternToBeFreed)
-  {
-    regfree(&entityIdPattern);
-    entityIdPatternToBeFreed = false;
-  }
-}
-
-
-
-/* ****************************************************************************
-*
-* EntityInfo::present - 
-*/
-void EntityInfo::present(const std::string& prefix)
-{
-  LM_F(("%sid:        %s", prefix.c_str(), entityId.c_str()));
-  LM_F(("%sisPattern: %s", prefix.c_str(), FT(isPattern)));
-  LM_F(("%stype:      %s", prefix.c_str(), entityType.c_str()));
-}
-
-
-
-/* ****************************************************************************
-*
-* MongoSubCache - 
-*/
-typedef struct MongoSubCache
-{
-  CachedSubscription* head;
-  CachedSubscription* tail;
-
-  // Statistics counters
-  int                 noOfRefreshes;
-  int                 noOfInserts;
-  int                 noOfRemoves;
-  int                 noOfUpdates;
-} MongoSubCache;
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCache - 
-*/
-MongoSubCache  mongoSubCache       = { NULL, NULL, 0 };
-bool           mongoSubCacheActive = false;
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheInit - 
-*/
-void mongoSubCacheInit(void)
-{
-  LM_T(LmtMongoSubCache, ("Initializing subscription cache"));
-
-  mongoSubCache.head   = NULL;
-  mongoSubCache.tail   = NULL;
-
-  mongoSubCacheStatisticsReset("mongoSubCacheInit");
-
-  mongoSubCacheActive = true;
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheItems - 
-*/
-int mongoSubCacheItems(void)
-{
-  CachedSubscription* cSubP = mongoSubCache.head;
-  int                 items = 0;
-
-  while (cSubP != NULL)
-  {
-    ++items;
-    cSubP = cSubP->next;
-  }
-
-  return items;
-}
-
-
-
-/* ****************************************************************************
-*
-* attributeMatch - 
-*/
-static bool attributeMatch(CachedSubscription* cSubP, const std::vector<std::string>& attrV)
-{
-  for (unsigned int ncvIx = 0; ncvIx < cSubP->notifyConditionVector.size(); ++ncvIx)
-  {
-    NotifyCondition* ncP = cSubP->notifyConditionVector[ncvIx];
-
-    for (unsigned int cvIx = 0; cvIx < ncP->condValueList.size(); ++cvIx)
-    {
-      for (unsigned int aIx = 0; aIx < attrV.size(); ++aIx)
-      {
-        if (ncP->condValueList[cvIx] == attrV[aIx])
-        {
-          return true;
-        }
-      }
-    }
-  }
-
-  return false;
-}
-
-
-
-/* ****************************************************************************
-*
-* servicePathMatch - 
-*/
-static bool servicePathMatch(CachedSubscription* cSubP, char* servicePath)
-{
-  const char*  spath;
-
-  // Special case. "/#" matches EVERYTHING
-  if ((strlen(servicePath) == 2) && (servicePath[0] == '/') && (servicePath[1] == '#') && (servicePath[2] == 0))
-  {
-    return true;
-  }
-
-  //
-  // Empty service path?
-  //
-  if ((strlen(servicePath) == 0) && (strlen(cSubP->servicePath) == 0))
-  {
-    return true;
-  }
-
-  //
-  // Default service path for a subscription is "/".
-  // So, if empty, set it to "/"
-  //
-  if (servicePath[0] == 0)
-  {
-    servicePath = (char*) "/";
-  }
-
-  spath = cSubP->servicePath;
-
-  //
-  // No wildcard - exact match
-  //
-  if (spath[strlen(spath) - 1] != '#')
-  {
-    return (strcmp(servicePath, cSubP->servicePath) == 0);
-  }
-
-
-  //
-  // Wildcard - remove '#' and compare to the length of _servicePath.
-  //            If equal upto the length of _servicePath, then it is a match
-  //
-  // Actually, there is more than one way to match here:
-  // If we have a servicePath of the subscription as 
-  // "/a/b/#", then the following service-paths must match:
-  //   1. /a/b
-  //   2. /a/b/ AND /a/b/.+  (any path below /a/b/)
-  //
-  // What should NOT match is "/a/b2".
-  //
-  unsigned int len = strlen(spath) - 2;
-
-  // 1. /a/b - (removing 2 chars from the cache-subscription removes "/#"
-
-  if ((spath[len] == '/') && (strlen(servicePath) == len) && (strncmp(spath, servicePath, len) == 0))
-  {
-    return true;
-  }
-
-  // 2. /a/b/.+
-  len = strlen(spath) - 1;
-  if (strncmp(spath, servicePath, len) == 0)
-  {
-    return true;
-  }
-  
-  return false;
-}
-
-
-
-/* ****************************************************************************
-*
-* subMatch - 
-*/
-static bool subMatch
-(
-  CachedSubscription*              cSubP,
-  const char*                      tenant,
-  const char*                      servicePath,
-  const char*                      entityId,
-  const char*                      entityType,
-  const std::vector<std::string>&  attrV
-)
-{
-  if ((cSubP->tenant == NULL) || (tenant == NULL) || (cSubP->tenant[0] == 0) || (tenant[0] == 0))
-  {
-    if ((cSubP->tenant != NULL) && (cSubP->tenant[0] != 0))
-    {
-      LM_T(LmtMongoSubCacheMatch, ("No match due to tenant I"));
-      return false;
-    }
-
-    if ((tenant != NULL) && (tenant[0] != 0))
-    {
-      LM_T(LmtMongoSubCacheMatch, ("No match due to tenant II"));
-      return false;
-    }
-  }
-  else if (strcmp(cSubP->tenant, tenant) != 0)
-  {
-    LM_T(LmtMongoSubCacheMatch, ("No match due to tenant III"));
-    return false;
-  }
-
-  if (servicePathMatch(cSubP, (char*) servicePath) == false)
-  {
-    LM_T(LmtMongoSubCacheMatch, ("No match due to servicePath"));
-    return false;
-  }
-
-
-  //
-  // If ONCHANGE and one of the attribute names in the scope vector
-  // of the subscription has the same name as the incoming attribute. there is a match.
-  //
-  if (!attributeMatch(cSubP, attrV))
-  {
-    LM_T(LmtMongoSubCacheMatch, ("No match due to attributes"));
-    return false;
-  }
-
-  for (unsigned int ix = 0; ix < cSubP->entityIdInfos.size(); ++ix)
-  {
-    EntityInfo* eiP = cSubP->entityIdInfos[ix];
-
-    if (eiP->match(entityId, entityType))
-    {
-      return true;
-    }
-  }
-
-  LM_T(LmtMongoSubCacheMatch, ("No match due to EntityInfo"));
-  return false;
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheMatch - 
-*/
-void mongoSubCacheMatch
-(
-  const char*                        tenant,
-  const char*                        servicePath,
-  const char*                        entityId,
-  const char*                        entityType,
-  const char*                        attr,
-  std::vector<CachedSubscription*>*  subVecP
-)
-{
-  CachedSubscription* cSubP = mongoSubCache.head;
-
-  while (cSubP != NULL)
-  {
-    std::vector<std::string> attrV;
-
-    attrV.push_back(attr);
-
-    if (subMatch(cSubP, tenant, servicePath, entityId, entityType, attrV))
-    {
-      subVecP->push_back(cSubP);
-      LM_T(LmtMongoSubCache, ("added subscription '%s': lastNotificationTime: %lu", cSubP->subscriptionId, cSubP->lastNotificationTime));
-    }
-
-    cSubP = cSubP->next;
-  }
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheMatch - 
-*/
-void mongoSubCacheMatch
-(
-  const char*                        tenant,
-  const char*                        servicePath,
-  const char*                        entityId,
-  const char*                        entityType,
-  const std::vector<std::string>&    attrV,
-  std::vector<CachedSubscription*>*  subVecP
-)
-{
-  CachedSubscription* cSubP = mongoSubCache.head;
-
-  while (cSubP != NULL)
-  {
-    if (subMatch(cSubP, tenant, servicePath, entityId, entityType, attrV))
-    {
-      subVecP->push_back(cSubP);
-      LM_T(LmtMongoSubCache, ("added subscription '%s': lastNotificationTime: %lu", cSubP->subscriptionId, cSubP->lastNotificationTime));
-    }
-
-    cSubP = cSubP->next;
-  }
-}
-
-
-
-/* ****************************************************************************
-*
-* cachedSubscriptionDestroy - 
-*/
-static void cachedSubscriptionDestroy(CachedSubscription* cSubP)
-{
-  if (cSubP->tenant != NULL)
-  {
-    free(cSubP->tenant);
-    cSubP->tenant = NULL;
-  }
-
-  if (cSubP->servicePath != NULL)
-  {
-    free(cSubP->servicePath);
-    cSubP->servicePath = NULL;
-  }
-
-  if (cSubP->subscriptionId != NULL)
-  {
-    free(cSubP->subscriptionId);
-    cSubP->subscriptionId = NULL;
-  }
-
-  if (cSubP->reference != NULL)
-  {
-    free(cSubP->reference);
-    cSubP->reference = NULL;
-  }
-
-  for (unsigned int ix = 0; ix < cSubP->entityIdInfos.size(); ++ix)
-  {
-    cSubP->entityIdInfos[ix]->release();
-    delete cSubP->entityIdInfos[ix];
-  }
-
-  cSubP->entityIdInfos.clear();
-
-  while (cSubP->attributes.size() > 0)
-  {
-    cSubP->attributes.erase(cSubP->attributes.begin());
-  }
-  cSubP->attributes.clear();
-
-  cSubP->notifyConditionVector.release();
-  cSubP->notifyConditionVector.vec.clear();
-
-  cSubP->next = NULL;
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheDestroy - 
-*/
-void mongoSubCacheDestroy(void)
-{
-  LM_T(LmtMongoSubCache, ("destroying subscription cache"));
-
-  CachedSubscription* cSubP  = mongoSubCache.head;
-
-  if (mongoSubCache.head == NULL)
-  {
-    return;
-  }
-
-  while (cSubP->next)
-  {
-    CachedSubscription* prev = cSubP;
-
-    cSubP = cSubP->next;
-    cachedSubscriptionDestroy(prev);
-    LM_T(LmtMongoSubCache,  ("removing CachedSubscription at %p", prev));
-    delete prev;
-  }
-
-  cachedSubscriptionDestroy(cSubP);
-  LM_T(LmtMongoSubCache,  ("removing CachedSubscription at %p", cSubP));
-  delete cSubP;
-
-  mongoSubCache.head  = NULL;
-  mongoSubCache.tail  = NULL;
-}
-
-
-
-/* ****************************************************************************
-*
-* tenantMatch - 
-*/
-static bool tenantMatch(const char* tenant1, const char* tenant2)
-{
-  //
-  // Removing complications with NULL, giving NULL tenants the value of an empty string;
-  //
-  tenant1 = (tenant1 == NULL)? "" : tenant1;
-  tenant2 = (tenant2 == NULL)? "" : tenant2;
-
-  if (strlen(tenant1) != strlen(tenant2))
-  {
-    return false;
-  }
-
-  if (strcmp(tenant1, tenant2) == 0)
-  {
-    return true;
-  }
-
-  return false;
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheItemLookup - 
-*
-* FIXME P7: lookups would be A LOT faster if the subCache used a hash-table instead of
-*           just a single linked link for the subscriptions.
-*           The hash could be adding each char in 'tenant' and 'subscriptionId' to a 8-bit
-*           integer and then we'd have a vector of 256 linked lists.
-*           Another way of implementing this would be to use a std::vector or std::map.
-*/
-CachedSubscription* mongoSubCacheItemLookup(const char* tenant, const char* subscriptionId)
-{
-  CachedSubscription* cSubP = mongoSubCache.head;
-
-  while (cSubP != NULL)
-  {
-    if ((tenantMatch(tenant, cSubP->tenant)) && (strcmp(subscriptionId, cSubP->subscriptionId) == 0))
-    {
-      return cSubP;
-    }
-
-    cSubP = cSubP->next;
-  }
-
-  return NULL;
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheUpdateStatisticsIncrement - 
-*/
-void mongoSubCacheUpdateStatisticsIncrement(void)
-{
-  ++mongoSubCache.noOfUpdates;
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheItemInsert - 
-*/
-static void mongoSubCacheItemInsert(CachedSubscription* cSubP)
-{
-  //
-  // First of all, insertion is done at the end of the list, so, 
-  // cSubP->next must ALWAYS be zero at insertion time
-  //
-  cSubP->next = NULL;
-
-  LM_T(LmtMongoSubCache, ("inserting sub '%s', lastNotificationTime: %lu", cSubP->subscriptionId, cSubP->lastNotificationTime));
-
-  ++mongoSubCache.noOfInserts;
-
-  // First insertion?
-  if ((mongoSubCache.head == NULL) && (mongoSubCache.tail == NULL))
-  {
-    mongoSubCache.head   = cSubP;
-    mongoSubCache.tail   = cSubP;
-
-    return;
-  }
-
-  mongoSubCache.tail->next  = cSubP;
-  mongoSubCache.tail        = cSubP;
-}
 
 
 
@@ -660,7 +74,7 @@ int mongoSubCacheItemInsert(const char* tenant, const BSONObj& sub)
   // 03. Create CachedSubscription
   //
   CachedSubscription* cSubP = new CachedSubscription();
-  LM_T(LmtMongoSubCache,  ("allocated CachedSubscription at %p", cSubP));
+  LM_T(LmtSubCache,  ("allocated CachedSubscription at %p", cSubP));
 
   if (cSubP == NULL)
   {
@@ -687,10 +101,10 @@ int mongoSubCacheItemInsert(const char* tenant, const BSONObj& sub)
   cSubP->throttling            = sub.hasField(CSUB_THROTTLING)?       getIntOrLongFieldAsLong(sub, CSUB_THROTTLING)       : -1;
   cSubP->expirationTime        = sub.hasField(CSUB_EXPIRATION)?       getIntOrLongFieldAsLong(sub, CSUB_EXPIRATION)       : 0;
   cSubP->lastNotificationTime  = sub.hasField(CSUB_LASTNOTIFICATION)? getIntOrLongFieldAsLong(sub, CSUB_LASTNOTIFICATION) : -1;
-  cSubP->count                 = sub.hasField(CSUB_COUNT)?            getIntOrLongFieldAsLong(sub, CSUB_COUNT)            : 0;
+  cSubP->count                 = 0;
   cSubP->next                  = NULL;
 
-  LM_T(LmtMongoSubCache, ("set lastNotificationTime to %lu for '%s' (from DB)", cSubP->lastNotificationTime, cSubP->subscriptionId));
+  LM_T(LmtSubCache, ("set lastNotificationTime to %lu for '%s' (from DB)", cSubP->lastNotificationTime, cSubP->subscriptionId));
 
   //
   // 05. Push Entity-data names to EntityInfo Vector (cSubP->entityInfos)
@@ -716,7 +130,7 @@ int mongoSubCacheItemInsert(const char* tenant, const BSONObj& sub)
   if (cSubP->entityIdInfos.size() == 0)
   {
     LM_E(("ERROR (no patterned entityId) - cleaning up"));
-    cachedSubscriptionDestroy(cSubP);
+    subCacheItemDestroy(cSubP);
     delete cSubP;
     return -3;
   }
@@ -766,12 +180,12 @@ int mongoSubCacheItemInsert(const char* tenant, const BSONObj& sub)
   if (cSubP->notifyConditionVector.size() == 0)  // Cleanup
   {
     LM_E(("ERROR (empty notifyConditionVector) - cleaning up"));
-    cachedSubscriptionDestroy(cSubP);
+    subCacheItemDestroy(cSubP);
     delete cSubP;
     return -4;
   }
 
-  mongoSubCacheItemInsert(cSubP);
+  subCacheItemInsert(cSubP);
 
   return 0;
 }
@@ -819,7 +233,7 @@ int mongoSubCacheItemInsert
   //
   CachedSubscription* cSubP = new CachedSubscription();
 
-  LM_T(LmtMongoSubCache,  ("allocated CachedSubscription at %p", cSubP));
+  LM_T(LmtSubCache,  ("allocated CachedSubscription at %p", cSubP));
 
   if (cSubP == NULL)
   {
@@ -856,7 +270,7 @@ int mongoSubCacheItemInsert
 
   if (cSubP->entityIdInfos.size() == 0)
   {
-    cachedSubscriptionDestroy(cSubP);
+    subCacheItemDestroy(cSubP);
     delete cSubP;
     return -4;
   }
@@ -887,9 +301,10 @@ int mongoSubCacheItemInsert
   cSubP->throttling            = sub.hasField(CSUB_THROTTLING)?       getIntOrLongFieldAsLong(sub, CSUB_THROTTLING) : -1;
   cSubP->expirationTime        = expirationTime;
   cSubP->lastNotificationTime  = lastNotificationTime;
+  cSubP->count                 = 0;
   cSubP->next                  = NULL;
 
-  LM_T(LmtMongoSubCache, ("set lastNotificationTime to %lu for '%s' (from DB)", cSubP->lastNotificationTime, cSubP->subscriptionId));
+  LM_T(LmtSubCache, ("set lastNotificationTime to %lu for '%s' (from DB)", cSubP->lastNotificationTime, cSubP->subscriptionId));
 
 
   //
@@ -935,265 +350,16 @@ int mongoSubCacheItemInsert
 
   if (cSubP->notifyConditionVector.size() == 0)
   {
-    cachedSubscriptionDestroy(cSubP);
+    subCacheItemDestroy(cSubP);
     delete cSubP;
     return -5;
   }
 
-  mongoSubCacheItemInsert(cSubP);
+  subCacheItemInsert(cSubP);
 
   return 0;
 }
 
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheItemInsert - create a new sub, fill it in, and add it to cache
-*/
-void mongoSubCacheItemInsert
-(
-  const char*               tenant,
-  const char*               servicePath,
-  SubscribeContextRequest*  scrP,
-  const char*               subscriptionId,
-  int64_t                   expirationTime,
-  int64_t                   throttling,
-  Format                    notifyFormat,
-  bool                      notificationDone,
-  int64_t                   lastNotificationTime
-)
-{
-  //
-  // Add the subscription to the subscription cache.
-  // But only if any of the entities in entityIdVector is pattern based -
-  // AND it is a subscription of type ONCHANGE
-  //
-
-  //
-  // 00. Check that the subscription is pattern based an ONCHANGE
-  //
-  bool onchange = false;
-
-  for (unsigned int ix = 0; ix < scrP->notifyConditionVector.size(); ++ix)
-  {
-    if (strcasecmp(scrP->notifyConditionVector[ix]->type.c_str(), ON_CHANGE_CONDITION) == 0)
-    {
-      onchange = true;
-      break;
-    }
-  }
-
-  if (!onchange)
-  {
-    return;
-  }
-
-    
-  //
-  // Now allocate the subscription
-  //
-  CachedSubscription* cSubP = new CachedSubscription();
-  LM_T(LmtMongoSubCache,  ("allocated CachedSubscription at %p", cSubP));
-
-
-  //
-  // 1. First the non-complex values
-  //
-  cSubP->tenant                = (tenant[0] == 0)? NULL : strdup(tenant);
-  cSubP->servicePath           = strdup(servicePath);
-  cSubP->subscriptionId        = strdup(subscriptionId);
-  cSubP->reference             = strdup(scrP->reference.get().c_str());
-  cSubP->expirationTime        = expirationTime;
-  cSubP->throttling            = throttling;
-  cSubP->lastNotificationTime  = lastNotificationTime;
-  cSubP->notifyFormat          = notifyFormat;
-  cSubP->next                  = NULL;
-  cSubP->count                 = (notificationDone == true)? 1 : 0;
-
-  LM_T(LmtMongoSubCache, ("inserting a new sub in cache (%s). lastNotifictionTime: %lu", cSubP->subscriptionId, cSubP->lastNotificationTime));
-
-
-  //
-  // 2. Then the values that have functions/methods for filling/parsing
-  //
-  cSubP->throttling = scrP->throttling.parse();
-  cSubP->notifyConditionVector.fill(scrP->notifyConditionVector);
-
-
-  //
-  // 3. Convert all EntityIds to EntityInfo
-  //
-  for (unsigned int ix = 0; ix < scrP->entityIdVector.vec.size(); ++ix)
-  {
-    EntityId* eIdP;
-
-    eIdP = scrP->entityIdVector.vec[ix];
-
-    EntityInfo* eP = new EntityInfo(eIdP->id, eIdP->type, eIdP->isPattern);
-    
-    cSubP->entityIdInfos.push_back(eP);
-  }
-
-
-  //
-  // 4. Insert all attributes 
-  //
-  for (unsigned int ix = 0; ix < scrP->attributeList.attributeV.size(); ++ix)
-  {
-    cSubP->attributes.push_back(scrP->attributeList.attributeV[ix]);
-  }
-
-
-  //
-  // 5. Now, insert the subscription in the cache
-  //
-  LM_T(LmtMongoSubCache, ("Inserting NEW sub '%s', lastNotificationTime: %lu", cSubP->subscriptionId, cSubP->lastNotificationTime));
-  mongoSubCacheItemInsert(cSubP);
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheStatisticsGet - 
-*/
-void mongoSubCacheStatisticsGet(int* refreshes, int* inserts, int* removes, int* updates, int* items, char* list, int listSize)
-{
-  *refreshes = mongoSubCache.noOfRefreshes;
-  *inserts   = mongoSubCache.noOfInserts;
-  *removes   = mongoSubCache.noOfRemoves;
-  *updates   = mongoSubCache.noOfUpdates;
-  *items     = mongoSubCacheItems();
-
-  CachedSubscription* cSubP = mongoSubCache.head;
-
-  //
-  // NOTE
-  //   If the listBuffer is not big enough to hold the entire list of cached subscriptions,
-  //   mongoSubCacheStatisticsGet returns the empty string and the list is excluded from
-  //   the response of "GET /statistics".
-  //   Minimum size of the list is set to 128 bytes and the needed size is detected later.
-  //   Again, if the list-buffer is not big enough to hold the entire text, no list is shown.
-  //
-  *list = 0;
-  if (listSize > 128)
-  {
-    while (cSubP != NULL)
-    {
-      char          msg[256];
-      unsigned int  bytesLeft = listSize - strlen(list);
-
-#if 0
-      //
-      // FIXME P5: To be removed once sub-update is OK in QA
-      //
-      int now = getCurrentTime();
-      snprintf(msg, sizeof(msg), "%s|N:%lu|E:%lu|T:%lu|DUR:%lu",
-               cSubP->subscriptionId,
-               cSubP->lastNotificationTime,
-               cSubP->expirationTime,
-               cSubP->throttling,
-               cSubP->expirationTime - now);
-#else
-        snprintf(msg, sizeof(msg), "%s", cSubP->subscriptionId);
-#endif
-      //
-      // If "msg" and ", " has no room in the "list", then no list is shown.
-      // (the '+ 2' if for the string ", ")
-      //
-      if (strlen(msg) + 2 > bytesLeft)
-      {
-        *list = 0;
-        break;
-      }
-      
-      if (list[0] != 0)
-        strcat(list, ", ");
-
-      strcat(list, msg);
-
-      cSubP = cSubP->next;
-    }
-  }
-
-  LM_T(LmtMongoSubCache, ("Statistics: refreshes: %d (%d)", *refreshes, mongoSubCache.noOfRefreshes));
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheStatisticsReset - 
-*/
-void mongoSubCacheStatisticsReset(const char* by)
-{
-  mongoSubCache.noOfRefreshes  = 0;
-  mongoSubCache.noOfInserts    = 0;
-  mongoSubCache.noOfRemoves    = 0;
-  mongoSubCache.noOfUpdates    = 0;
-
-  LM_T(LmtMongoSubCache, ("Statistics reset by %s: refreshes is now %d", by, mongoSubCache.noOfRefreshes));
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCachePresent - 
-*/
-void mongoSubCachePresent(const char* title)
-{
-  CachedSubscription* cSubP = mongoSubCache.head;
-
-  LM_T(LmtMongoSubCache, ("----------- %s ------------", title));
-
-  while (cSubP != NULL)
-  {
-    LM_T(LmtMongoSubCache, ("o %s (tenant: %s, LNT: %lu, THR: %d)", cSubP->subscriptionId, cSubP->tenant, cSubP->lastNotificationTime, cSubP->throttling));
-    cSubP = cSubP->next;
-  }
-
-  LM_T(LmtMongoSubCache, ("--------------------------------"));
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheItemRemove - 
-*/
-int mongoSubCacheItemRemove(CachedSubscription* cSubP)
-{
-  CachedSubscription* current = mongoSubCache.head;
-  CachedSubscription* prev    = NULL;
-
-  LM_T(LmtMongoSubCache, ("in mongoSubCacheItemRemove, trying to remove '%s'", cSubP->subscriptionId));
-  while (current != NULL)
-  {
-    if (current == cSubP)
-    {
-      if (cSubP == mongoSubCache.head)  { mongoSubCache.head = cSubP->next; }
-      if (cSubP == mongoSubCache.tail)  { mongoSubCache.tail = prev;        }
-      if (prev != NULL)                 { prev->next         = cSubP->next; }
-
-      LM_T(LmtMongoSubCache, ("in mongoSubCacheItemRemove, REMOVING '%s'", cSubP->subscriptionId));
-      ++mongoSubCache.noOfRemoves;
-
-      cachedSubscriptionDestroy(cSubP);
-      delete cSubP;
-
-      return 0;
-    }
-
-    prev    = current;
-    current = current->next;
-  }
-
-  LM_E(("Runtime Error (item to remove from sub-cache not found)"));
-  return -1;
-}
 
 
 
@@ -1211,9 +377,9 @@ int mongoSubCacheItemRemove(CachedSubscription* cSubP)
 *
 *   I.e. the subscriptions is for ONCHANGE.
 */
-static void mongoSubCacheRefresh(const std::string& database)
+void mongoSubCacheRefresh(const std::string& database)
 {
-  LM_T(LmtMongoSubCache, ("Refreshing subscription cache for DB '%s'", database.c_str()));
+  LM_T(LmtSubCache, ("Refreshing subscription cache for DB '%s'", database.c_str()));
 
   BSONObj                   query       = BSON("conditions.type" << "ONCHANGE");
   std::string               db          = database;
@@ -1228,7 +394,6 @@ static void mongoSubCacheRefresh(const std::string& database)
     return;
   }
 
-  // Call the treat function for each subscription
   int subNo = 0;
   while (moreSafe(cursor))
   {
@@ -1249,259 +414,43 @@ static void mongoSubCacheRefresh(const std::string& database)
     }
   }
 
-  LM_T(LmtMongoSubCache, ("Added %d subscriptions for database '%s'", subNo, database.c_str()));
+  LM_T(LmtSubCache, ("Added %d subscriptions for database '%s'", subNo, database.c_str()));
 }
 
 
 
 /* ****************************************************************************
 *
-* mongoSubCacheRefresh - 
-*
-* WARNING
-*  The cache semaphore must be taken before this function is called:
-*    cacheSemTake(__FUNCTION__, "Reason");
-*  And released after mongoSubCacheRefresh finishes, of course.
+* mongoSubCacheUpdate - update subscription in mongo with count and lastNotificationTime
 */
-void mongoSubCacheRefresh(void)
+void mongoSubCacheUpdate(const std::string& tenant, const std::string& subId, long long count, long long lastNotificationTime)
 {
-  std::vector<std::string> databases;
+  std::string  collection  = getSubscribeContextCollectionName(tenant);
+  BSONObj      condition;
+  BSONObj      update;
+  std::string  err;
 
-  LM_T(LmtMongoSubCache, ("Refreshing subscription cache"));
-
-  // Empty the cache
-  mongoSubCacheDestroy();
-
-  // Get list of database
-  getOrionDatabases(databases);
-
-  // Add the 'default tenant'
-  databases.push_back(getDbPrefix());
-
-
-  // Now refresh the subCache for each and every tenant
-  for (unsigned int ix = 0; ix < databases.size(); ++ix)
+  if (count != 0)
   {
-    LM_T(LmtMongoSubCache, ("DB %d: %s", ix, databases[ix].c_str()));
-    mongoSubCacheRefresh(databases[ix]);
-  }
-
-  ++mongoSubCache.noOfRefreshes;
-  LM_T(LmtMongoSubCache, ("Refreshed subscription cache [%d]", mongoSubCache.noOfRefreshes));
-
-  //
-  // FIXME P5: Remove this debugging code once the refresh works again
-  //
-#if 0
-
-  //
-  // DEBUG STARTS HERE
-  //
-  CachedSubscription* cSubP = mongoSubCache.head;
-  int                 items = 0;
-
-  while (cSubP != NULL)
-  {
-    LM_T(LmtMongoSubCache, ("%s: TEN:%s, SPath:%s, THR:%lu, EXP:%lu, LNT:%lu REF:%s, Es:%d, As:%d, NCs:%d",
-                            cSubP->subscriptionId,
-                            cSubP->tenant,
-                            cSubP->servicePath,
-                            cSubP->throttling,
-                            cSubP->expirationTime,
-                            cSubP->lastNotificationTime,
-                            cSubP->reference,
-                            cSubP->entityIdInfos.size(),
-                            cSubP->attributes.size(),
-                            cSubP->notifyConditionVector.size()));
-
-    cSubP = cSubP->next;
-    ++items;
-  }
-
-  LM_T(LmtMongoSubCache, ("%d subs in sub.cache", items));
-
-  //
-  // DEBUG ENDS HERE
-  //
-#endif
-}
-
-
-
-/* ****************************************************************************
-*
-* CachedSubSaved - 
-*/
-typedef struct CachedSubSaved
-{
-  long long  lastNotificationTime;
-  long long  count;
-} CachedSubSaved;
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheSync - 
-*
-* 1. Save subscriptionId, lastNotificationTime, and count for all items in cache (savedSubV)
-* 2. Refresh cache (count set to 0)
-* 3. Compare lastNotificationTime in savedSubV with the new cache-contents and:
-*    3.1 Update cache-items where 'saved lastNotificationTime' > 'cached lastNotificationTime'
-*    3.2 Remember this more correct lastNotificationTime (must be flushed to mongo) -
-*        by clearing out (set to 0) those lastNotificationTimes that are newer in cache
-* 4. Update 'count' for each item in savedSubV where count != 0
-* 5. Update 'lastNotificationTime' foreach item in savedSubV where lastNotificationTime != 0
-* 6. Free the vector created in step 1 - savedSubV
-*/
-void mongoSubCacheSync(void)
-{
-  std::map<std::string, CachedSubSaved*> savedSubV;
-
-  cacheSemTake(__FUNCTION__, "Synchronizing subscription cache");
-
-
-  //
-  // 1. Save subscriptionId, lastNotificationTime, and count for all items in cache
-  //
-  CachedSubscription* cSubP = mongoSubCache.head;
-
-  while (cSubP != NULL)
-  {
-    CachedSubSaved* cssP       = new CachedSubSaved();
-
-    cssP->lastNotificationTime = cSubP->lastNotificationTime;
-    cssP->count                = cSubP->count;
-
-    savedSubV[cSubP->subscriptionId]= cssP;
-    cSubP = cSubP->next;
-  }
-
-  LM_T(LmtMongoCacheSync, ("Pushed back %d items to savedSubV", savedSubV.size()));
-
-
-  //
-  // 2. Refresh cache (count set to 0)
-  //
-  mongoSubCacheRefresh();
-
-
-  //
-  // 3. Compare lastNotificationTime in savedSubV with the new cache-contents
-  //
-  cSubP = mongoSubCache.head;
-  while (cSubP != NULL)
-  {
-    CachedSubSaved* cssP = savedSubV[cSubP->subscriptionId];
-    if ((cssP != NULL) && (cssP->lastNotificationTime <= cSubP->lastNotificationTime))
-    {
-      // cssP->lastNotificationTime older than what's currently in DB => throw away
-      cssP->lastNotificationTime = 0;
-    }
-    cSubP = cSubP->next;
-  }
-
-
-  //
-  // 4. Update 'count' for each item in savedSubV where count != 0
-  // 5. Update 'lastNotificationTime' foreach item in savedSubV where lastNotificationTime != 0
-  //
-  cSubP = mongoSubCache.head;
-  while (cSubP != NULL)
-  {
-    CachedSubSaved* cssP = savedSubV[cSubP->subscriptionId];
-
-    if (cssP == NULL)
-    {
-      cSubP = cSubP->next;
-      continue;
-    }
-
-    std::string  tenant      = (cSubP->tenant == NULL)? "" : cSubP->tenant;
-    std::string  collection  = getSubscribeContextCollectionName(tenant);
-    BSONObj      condition;
-    BSONObj      update;
-    std::string  err;
-
-    if (cssP->count != 0)
-    {
-      // Update count
-      condition = BSON("_id" << OID(cSubP->subscriptionId));
-      update    = BSON("$inc" << BSON(CSUB_COUNT << cssP->count));
-    }
-    else if (cssP->lastNotificationTime != 0)
-    {
-      // Update lastNotificationTime
-      condition = BSON("_id" << OID(cSubP->subscriptionId) << CSUB_LASTNOTIFICATION << BSON("$lt" << cssP->lastNotificationTime));
-      update    = BSON("$set" << BSON(CSUB_LASTNOTIFICATION << cssP->lastNotificationTime));
-    }
-    else
-    {
-      cSubP = cSubP->next;
-      continue;
-    }
+    // Update count
+    condition = BSON("_id"  << OID(subId));
+    update    = BSON("$inc" << BSON(CSUB_COUNT << count));
 
     if (collectionUpdate(collection, condition, update, false, &err) != true)
     {
-      LM_E(("Internal Error (error updating 'count' and 'lastNotification' for a subscription)"));
+      LM_E(("Internal Error (error updating 'count' for a subscription)"));
     }
-
-    cSubP = cSubP->next;
   }
 
-
-  //
-  // 6. Free the vector savedSubV
-  //
-  for (std::map<std::string, CachedSubSaved*>::iterator it = savedSubV.begin(); it != savedSubV.end(); ++it)
+  if (lastNotificationTime != 0)
   {
-    delete it->second;
+    // Update lastNotificationTime
+    condition = BSON("_id" << OID(subId) << CSUB_LASTNOTIFICATION << BSON("$lt" << lastNotificationTime));
+    update    = BSON("$set" << BSON(CSUB_LASTNOTIFICATION << lastNotificationTime));
+
+    if (collectionUpdate(collection, condition, update, false, &err) != true)
+    {
+      LM_E(("Internal Error (error updating 'lastNotification' for a subscription)"));
+    }
   }
-  savedSubV.clear();
-
-
-  cacheSemGive(__FUNCTION__, "Synchronizing subscription cache");
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheRefresherThread - 
-*/
-static void* mongoSubCacheRefresherThread(void* vP)
-{
-  extern int subCacheInterval;
-
-  while (1)
-  {
-    sleep(subCacheInterval);
-    mongoSubCacheSync();
-  }
-
-  return NULL;
-}
-
-
-
-/* ****************************************************************************
-*
-* mongoSubCacheStart - 
-*/
-void mongoSubCacheStart(void)
-{
-  pthread_t  tid;
-  int        ret;
-
-  // Populate subscription cache from database
-  mongoSubCacheRefresh();
-
-  ret = pthread_create(&tid, NULL, mongoSubCacheRefresherThread, NULL);
-
-  if (ret != 0)
-  {
-    LM_E(("Runtime Error (error creating thread: %d)", ret));
-    return;
-  }
-  pthread_detach(tid);
 }
