+++ src/lib/mongoBackend/mongoSubscribeContext.cpp
@@ -33,7 +33,7 @@
 #include "mongoBackend/dbConstants.h"
 #include "mongoBackend/mongoSubscribeContext.h"
 #include "mongoBackend/connectionOperations.h"
-#include "mongoBackend/mongoSubCache.h"
+#include "cache/subCache.h"
 #include "ngsi10/SubscribeContextRequest.h"
 #include "ngsi10/SubscribeContextResponse.h"
 #include "ngsi/StatusCode.h"
@@ -168,18 +168,18 @@ HttpStatusCode mongoSubscribeContext
     //
     std::string oidString = oid.toString();
 
-    LM_T(LmtMongoSubCache, ("inserting a new sub in cache (%s)", oidString.c_str()));
+    LM_T(LmtSubCache, ("inserting a new sub in cache (%s)", oidString.c_str()));
 
     cacheSemTake(__FUNCTION__, "Inserting subscription in cache");
-    mongoSubCacheItemInsert(tenant.c_str(),
-                            servicePath.c_str(),
-                            requestP,
-                            oidString.c_str(),
-                            expiration,
-                            throttling,
-                            notifyFormat,
-                            notificationDone,
-                            lastNotificationTime);
+    subCacheItemInsert(tenant.c_str(),
+                       servicePath.c_str(),
+                       requestP,
+                       oidString.c_str(),
+                       expiration,
+                       throttling,
+                       notifyFormat,
+                       notificationDone,
+                       lastNotificationTime);
     cacheSemGive(__FUNCTION__, "Inserting subscription in cache");
 
     reqSemGive(__FUNCTION__, "ngsi10 subscribe request", reqSemTaken);
