+++ test/unittests/CMakeLists.txt
@@ -263,6 +263,7 @@ SET (STATIC_LIBS
     ngsi9
     ngsi10
     ngsi
+    cache
     mongoBackend
     orionTypes
     common
