+++ test/unittests/mongoBackend/mongoUpdateContext_withOnchangeSubscriptions_test.cpp
@@ -32,7 +32,7 @@
 #include "common/globals.h"
 #include "mongoBackend/MongoGlobal.h"
 #include "mongoBackend/mongoUpdateContext.h"
-#include "mongoBackend/mongoSubCache.h"
+#include "cache/subCache.h"
 #include "ngsi/EntityId.h"
 #include "ngsi/ContextElementResponse.h"
 #include "ngsi10/UpdateContextRequest.h"
@@ -295,8 +295,8 @@ static void prepareDatabase(bool useSubCache = true)
   /* Given that preparation including csubs, we have to init cache */
   if (useSubCache == true)
   {
-    mongoSubCacheInit();
-    mongoSubCacheRefresh();
+    subCacheInit();
+    subCacheRefresh();
   }
 }
 
@@ -378,8 +378,8 @@ static void prepareDatabaseWithNoTypeSubscriptions(void)
     connection->insert(SUBSCRIBECONTEXT_COLL, sub5);
 
     /* Given that preparation including csubs, we have to initialize the subscription cache */
-    mongoSubCacheInit();
-    mongoSubCacheRefresh();
+    subCacheInit();
+    subCacheRefresh();
 }
 
 /* ****************************************************************************
@@ -435,7 +435,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_updateMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
 
@@ -499,7 +499,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_appendMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
 
@@ -558,7 +558,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_deleteMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
 
@@ -625,8 +625,8 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_updateMatch_noType)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* sub1P = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
-    CachedSubscription* sub2P = mongoSubCacheItemLookup("", "51307b66f481db11bf860004");
+    CachedSubscription* sub1P = subCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* sub2P = subCacheItemLookup("", "51307b66f481db11bf860004");
 
     ASSERT_TRUE(sub1P != NULL);
     ASSERT_TRUE(sub2P != NULL);
@@ -700,8 +700,8 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_appendMatch_noType)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* sub1P = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
-    CachedSubscription* sub2P = mongoSubCacheItemLookup("", "51307b66f481db11bf860004");
+    CachedSubscription* sub1P = subCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* sub2P = subCacheItemLookup("", "51307b66f481db11bf860004");
 
     ASSERT_TRUE(sub1P != NULL);
     ASSERT_TRUE(sub2P != NULL);
@@ -771,8 +771,8 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_deleteMatch_noType)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* sub1P = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
-    CachedSubscription* sub2P = mongoSubCacheItemLookup("", "51307b66f481db11bf860004");
+    CachedSubscription* sub1P = subCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* sub2P = subCacheItemLookup("", "51307b66f481db11bf860004");
 
     ASSERT_TRUE(sub1P != NULL);
     ASSERT_TRUE(sub2P != NULL);
@@ -838,7 +838,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_updateMatch_pattern)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860003");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860003");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -904,7 +904,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_appendMatch_pattern)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860003");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860003");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -966,7 +966,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_deleteMatch_pattern)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860003");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860003");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1029,7 +1029,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_updateMatch_pattern_noT
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860005");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860005");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1094,7 +1094,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_appendMatch_pattern_noT
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860005");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860005");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1155,7 +1155,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_deleteMatch_pattern_noT
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860005");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860005");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1218,7 +1218,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_updateMatchDisjoint)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1281,7 +1281,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_appendMatchDisjoint)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1344,7 +1344,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_deleteMatchDisjoint)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1396,7 +1396,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_updateNoMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(20000000, subP->lastNotificationTime);
@@ -1614,7 +1614,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_updateMixMatchNoMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1682,7 +1682,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_appendMixMatchNoMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1746,7 +1746,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_deleteMixMatchNoMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1812,7 +1812,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_update2Matches1Notifica
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1881,7 +1881,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_append2Matches1Notifica
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -1944,7 +1944,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, Cond1_delete2Matches1Notifica
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2009,7 +2009,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_updateMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2075,7 +2075,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_appendMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2137,7 +2137,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_deleteMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2201,7 +2201,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_updateMatchDisjoint)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2265,7 +2265,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_appendMatchDisjoint)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2329,7 +2329,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_deleteMatchDisjoint)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2599,7 +2599,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_updateMixMatchNoMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2667,7 +2667,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_appendMixMatchNoMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2731,7 +2731,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_deleteMixMatchNoMatch)
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2797,7 +2797,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_update2Matches1Notifica
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2866,7 +2866,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_append2Matches1Notifica
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -2929,7 +2929,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, CondN_delete2Matches1Notifica
     EXPECT_EQ(SccOk, ms);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860002");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860002");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(1360232700, subP->lastNotificationTime);
@@ -3006,7 +3006,7 @@ TEST(mongoUpdateContext_withOnchangeSubscriptions, DISABLED_MongoDbQueryFail)
     EXPECT_EQ("", RES_CER_STATUS(0).details);
 
     /* Check lastNotification */
-    CachedSubscription* subP = mongoSubCacheItemLookup("", "51307b66f481db11bf860001");
+    CachedSubscription* subP = subCacheItemLookup("", "51307b66f481db11bf860001");
 
     ASSERT_TRUE(subP != NULL);
     EXPECT_EQ(20000000, subP->lastNotificationTime);
