+++ test/unittests/unittest.cpp
@@ -26,7 +26,7 @@
 
 #include "unittest.h"
 #include "testInit.h"
-#include "mongoBackend/mongoSubCache.h"
+#include "cache/subCache.h"
 
 
 
@@ -136,8 +136,8 @@ void utInit(void)
   //
   servicePathVector.clear();
 
-  // Init subs cache (this initialization is overridden in test that use csubs)
-  mongoSubCacheInit();
+  // Init subs cache (this initialization is overridden in tests that use csubs)
+  subCacheInit();
 }
